requirements.txt
torch>=1.7.1
segment-anything>=1.0.0
opencv-python>=4.5.5
Pillow>=9.0.1
numpy>=1.21.0
torchvision>=0.20.1
fastapi>=0.115.5
uvicorn>=0.32.1
python-multipart>=0.0.18
scikit-learn>=1.5.2
